+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix #1724: createOrReplace function does not work properly for Custom defined resources
 * Fix #1789: Create or replace on operation seems broken
 * Fix #1782: Informer Deadlock; Fix lock typo in SharedProcessor
+* Fix #1607: WaitUntilReady for lists
 
 #### Improvements
 * Fix #1797: Utils.waitUntilReady should record the stack trace of the caller before rethrowing an exception